+++ module/post-fs-data.sh
@@ -3,11 +3,6 @@ if magisk --denylist status; then
     magisk --denylist rm com.google.android.gms
 fi
 
-# Remove conflicting modules if installed
-if [ -d /data/adb/modules/safetynet-fix ]; then
-    touch /data/adb/modules/safetynet-fix/remove
-fi
-
 # Conditional early sensitive properties
 
 resetprop_if_diff() {